requirements.txt
deepeval
requests
python-dotenv
pydantic
pytest